changed helper methods for user per devise

+++ app/controllers/posts_controller.rb
@@ -4,12 +4,15 @@ class PostsController < ApplicationController
   end
 
   def new
-  	require_user
+    if !user_signed_in?
+      redirect_to user_session_path
+    end
   end
 
   def create
     user = current_user
     post = user.posts.new(post_params)
+    binding.pry
     if post.save
       redirect_to root_path
     else
@@ -19,12 +22,16 @@ class PostsController < ApplicationController
   end
 
   def edit
-    require_user
+    if !user_signed_in?
+      redirect_to user_session_path
+    end
     @post = Post.find(params[:id])
   end
 
   def update
-    require_user
+    if !user_signed_in?
+      redirect_to user_session_path
+    end
     @post = Post.find(params[:id])
     if @post.update_attributes(post_params)
       redirect_to root_path
@@ -35,11 +42,13 @@ class PostsController < ApplicationController
   end
 
   def destroy
-    require_user
+    if !user_signed_in?
+      redirect_to user_session_path
+    end
     post = Post.find(params[:id])
-    if logged_in? && current_user.id == post.user_id   
-    post.destroy
-    redirect_to posts_path
+    if user_signed_in? && current_user.id == post.user_id   
+      post.destroy
+      redirect_to posts_path
     end
   end
